+++ apple-pay/index.js
@@ -13,7 +13,7 @@ var createAssetsUrl = require('../lib/create-assets-url');
 var createDeferredClient = require('../lib/create-deferred-client');
 var Promise = require('../lib/promise');
 var errors = require('./errors');
-var VERSION = "3.82.0";
+var VERSION = "3.85.2";
 var wrapPromise = require('@braintree/wrap-promise');
 
 /**