+++ payment-request/index.js
@@ -11,7 +11,7 @@ var basicComponentVerification = require('../lib/basic-component-verification');
 var createDeferredClient = require('../lib/create-deferred-client');
 var createAssetsUrl = require('../lib/create-assets-url');
 var wrapPromise = require('@braintree/wrap-promise');
-var VERSION = "3.82.0";
+var VERSION = "3.85.2";
 
 /**
  * @static