+++ three-d-secure/index.js
@@ -9,7 +9,7 @@ var createAssetsUrl = require('../lib/create-assets-url');
 var BraintreeError = require('../lib/braintree-error');
 var analytics = require('../lib/analytics');
 var errors = require('./shared/errors');
-var VERSION = "3.82.0";
+var VERSION = "3.85.2";
 var Promise = require('../lib/promise');
 var wrapPromise = require('@braintree/wrap-promise');