+++ visa-checkout/index.js
@@ -12,7 +12,7 @@ var createAssetsUrl = require('../lib/create-assets-url');
 var VisaCheckout = require('./visa-checkout');
 var analytics = require('../lib/analytics');
 var errors = require('./errors');
-var VERSION = "3.82.0";
+var VERSION = "3.85.2";
 var Promise = require('../lib/promise');
 var wrapPromise = require('@braintree/wrap-promise');