+++ CHANGELOG.md
@@ -1,5 +1,54 @@
 # CHANGELOG
 
+# 3.85.2
+
+  - Venmo
+    - Fix issue where iOS Chrome was reporting as a supported browser when Venmo was configured for desktop
+    - Improve Venmo modal UX in desktop flow
+  - UnionPay
+    - Fix typo in our GraphQL Tokenization CREDIT_CARD_BRAND_MAP
+
+# 3.85.1
+
+  - Venmo
+    - Fix issue where iOS Chrome was reporting as a supported browser
+
+# 3.85.0 
+
+ - Client
+      - Add Elo, Hiper, and Hipercard graphQL adapters to card
+        tokenization responses
+  - Hosted Fields
+      - Add support for Elo, Hiper, and Hipercard in tokenization payload
+      - Fix issue where Hosted Fields won't lose focus if scrolled out of view on iOS
+  - Payment Request
+      - Add support for Elo, Hiper, and Hipercard in tokenization payload
+  - Venmo
+      - Expose `paymentContextId` when available
+
+# 3.84.0
+
+  - Client
+      - Add Elo, Hiper, and Hipercard to `supportedCardTypes`
+      - Fix issue where analytics event would report redundant errors
+        when client fails to set up (\#606)
+  - Venmo
+      - Adjust UI for better navigation
+
+# 3.83.0
+
+  - Data Collector
+      - Updates `clientMetadataId` to `riskCorrelationId`
+        (`clientMetadataId` is treated as an alias)
+  - Hosted Fields
+      - Allow passing `iframeTitle` in fields configuration to customize
+        iframe titles for field (\#545)
+  - Local Payments
+      - Add `options.displayName` to `startPayment`
+  - Venmo
+      - Fix issue where Facebook on Android would report as a supported
+        browser
+
 # 3.82.0
 
   - Update @braintree/browser-detection to v1.12.1